fix(buzz): move the label helper to utils/buzz and close two test bypasses
buzzTransactionLabel now lives beside parseBuzzTransactionDetails and
getAccountTypeLabel in utils/buzz.ts. The other transaction renderer,
/user/transactions, already imports from there; the next person changing
label copy would not have found a helper inside components/Buzz/Dashboard.

The source pin on the dashboard asserted one spelling of the call rather
than the property, so keeping the call behind a condition
(`cond ? buzzTransactionLabel(t) : t.description`) passed. It now asserts
the component never touches transaction.description at all, and matches the
call through whitespace so a prettier wrap is not a false red.

The jargon predicate hand-listed leg names case-sensitively, so 'holdfee',
'Escrow' and 'placementId 76' all slipped through. Leg names are derived
from the ledger table's own keys, matched case-insensitively, and the id
shape is loosened; `forfeit` is excluded because it is also an English word
and "Forfeited sticker placement" is copy, not jargon.

Also assert the fallback for LicenseFee, and pin the spoof labels as
literals rather than deriving them from the enum.
