+++ src/components/Buzz/Dashboard/BuzzDashboardOverview.tsx
@@ -27,8 +27,7 @@ import { DaysFromNow } from '~/components/Dates/DaysFromNow';
 import { UserBuzz } from '~/components/User/UserBuzz';
 import { BuzzTopUpCard } from '~/components/Buzz/BuzzTopUpCard';
 import { formatDate } from '~/utils/date-helpers';
-import { buzzTransactionLabel } from '~/components/Buzz/Dashboard/transaction-description';
-import { getAccountTypeLabel } from '~/utils/buzz';
+import { buzzTransactionLabel, getAccountTypeLabel } from '~/utils/buzz';
 import { useBuzzCurrencyConfig } from '~/components/Currency/useCurrencyConfig';
 import { hexToRgbOpenEnded } from '~/utils/mantine-css-helpers';
 import classes from '~/components/Buzz/buzz.module.scss';
