+++ src/server/services/__tests__/placement-escrow.service.test.ts
@@ -20,6 +20,7 @@ vi.mock('~/server/services/buzz.service', () => ({
   createBuzzTransaction,
 }));
 
+
 // Wholesale on purpose: the real module loads `base.reward`, which builds a
 // ClickHouse client, redis handles and prom collectors at import time. The
 // reward's own behaviour is tested in stickerPlacementAccepted.reward.test.ts;
@@ -93,6 +94,7 @@ const legKey = (placementId: number, kind: string) => `${placementId}:${kind}`;
 
 const queryRaw = vi.fn(async () => [] as { id: number }[]);
 const txCommits: string[][] = [];
+
 
 Object.assign(dbWriteMock, {
   // Interactive transaction: the callback gets the same client. Good enough to
