+++ src/utils/__tests__/buzz-transaction-label.test.ts
@@ -1,9 +1,9 @@
 import { readFileSync } from 'fs';
 import path from 'path';
 import { describe, expect, it } from 'vitest';
-import { buzzTransactionLabel } from '~/components/Buzz/Dashboard/transaction-description';
 import { TransactionType } from '~/shared/constants/buzz.constants';
 import { PLACEMENT_LEDGER_TEXT } from '~/shared/utils/placement';
+import { buzzTransactionLabel } from '~/utils/buzz';
 
 /**
  * Pre-#4212 rows still in prod: 1,527 as of 2026-09-09. They are why the
@@ -20,6 +20,21 @@ const LEDGER_DESCRIPTIONS = Object.values(PLACEMENT_LEDGER_TEXT).flatMap((byKind
   Object.values(byKind)
 );
 
+/**
+ * Derived from the leg keys rather than hand-listed, so a leg added to
+ * `PlacementTransactionKind` is covered without anyone remembering to widen this.
+ * camelCase keys only — `forfeit` is also an English word, and matching it would
+ * fail on "Forfeited sticker placement", which is copy rather than jargon. A
+ * placement id in such a string is still caught by `PLACEMENT_ID`.
+ */
+const LEG_NAMES = new RegExp(
+  Object.keys(PLACEMENT_LEDGER_TEXT.sticker)
+    .filter((kind) => /[A-Z]/.test(kind))
+    .join('|'),
+  'i'
+);
+const PLACEMENT_ID = /placement\s*(id\s*)?#?\d+/i;
+
 describe('buzzTransactionLabel', () => {
   it.each(LEGACY_PLACEMENT_DESCRIPTIONS)('hides the legacy description %s', (description) => {
     expect(buzzTransactionLabel({ type: TransactionType.Fee, description })).toBe('Fee');
@@ -35,10 +50,8 @@ describe('buzzTransactionLabel', () => {
    * question the table can answer wrongly.
    */
   it.each(LEDGER_DESCRIPTIONS)('%s carries no internal leg name or placement id', (description) => {
-    expect(description).not.toMatch(/placement \d+/i);
-    expect(description).not.toMatch(
-      /holdFee|holdPrincipal|toOwner|feeToOwner|toSeller|principalToPlacer|feeToPlacer|escrow/
-    );
+    expect(description).not.toMatch(PLACEMENT_ID);
+    expect(description).not.toMatch(LEG_NAMES);
   });
 
   // Literal pins for three of the eighteen, so a copy change is a red diff
@@ -70,16 +83,19 @@ describe('buzzTransactionLabel', () => {
   /**
    * A tip's description is written by the tipper, so an allowlist checked on any
    * type would sell a system-voice line in someone else's ledger for the price of
-   * one tip.
+   * one tip. Expected labels are literals: deriving them from the enum would pass
+   * for a reason unrelated to the guard.
    */
-  it.each([TransactionType.Tip, TransactionType.Refund, TransactionType.Bounty])(
-    'does not render an allowlisted string carried by type %s',
-    (type) => {
-      expect(
-        buzzTransactionLabel({ type, description: 'Someone placed a sticker on your image' })
-      ).toBe(TransactionType[type]);
-    }
-  );
+  it.each([
+    [TransactionType.Tip, 'Tip'],
+    [TransactionType.Refund, 'Refund'],
+    [TransactionType.Bounty, 'Bounty'],
+    [TransactionType.LicenseFee, 'License Fee'],
+  ])('does not render an allowlisted string carried by type %s', (type, label) => {
+    expect(
+      buzzTransactionLabel({ type, description: 'Someone placed a sticker on your image' })
+    ).toBe(label);
+  });
 
   it('still renders any description on the types that carry only user-facing copy', () => {
     expect(
@@ -107,17 +123,21 @@ describe('declined-fee copy reads as money kept, not money charged', () => {
 
 /**
  * The helper is pure, so every test above passes against a dashboard that never
- * calls it. This is the only thing tying the fix to the surface the ticket is
- * about.
+ * calls it. A text scan cannot see rendering — what it can see is the component
+ * holding the raw field at all, which is the shape of every way the description
+ * gets out unfiltered.
  */
 describe('the dashboard renders through the helper', () => {
-  const source = readFileSync(path.resolve(__dirname, '../BuzzDashboardOverview.tsx'), 'utf-8');
+  const source = readFileSync(
+    path.resolve(__dirname, '../../components/Buzz/Dashboard/BuzzDashboardOverview.tsx'),
+    'utf-8'
+  );
 
   it('calls buzzTransactionLabel', () => {
-    expect(source).toContain('buzzTransactionLabel(transaction)');
+    expect(source).toMatch(/buzzTransactionLabel\(\s*transaction\s*\)/);
   });
 
-  it('renders no description of its own', () => {
-    expect(source).not.toMatch(/\{transaction\.description\}|transaction\.description \?/);
+  it('never touches transaction.description itself', () => {
+    expect(source).not.toMatch(/transaction\.description/);
   });
 });
